+++ apps/meteor/app/cloud/server/functions/saveRegistrationData.ts
@@ -1,6 +1,6 @@
 import { Settings } from '@rocket.chat/models';
 
-import { applyLicense } from '../../../../ee/app/license/server/startup';
+import { applyLicense } from '../../../../ee/app/license/server/applyLicense';
 import { settings } from '../../../settings/server';
 import { syncCloudData } from './syncWorkspace/syncCloudData';
